Extract XInclude tests to a separate class
