+++ intellij-plugin-structure/tests/src/test/kotlin/com/jetbrains/plugin/structure/mocks/MockPluginsTest.kt
@@ -625,135 +625,6 @@ class MockPluginsTest(fileSystemType: FileSystemType) : BasePluginManagerTest<Id
     assertTrue(plugin.hasDotNetPart)
   }
 
-  @Test
-  fun `xinclude includeUnless with system property being set to false`() {
-    withSystemProperty("com.jetbrains.plugins.structure.mocks.optionalDependency", false) {
-      val plugin = buildPluginSuccess(emptyList()) {
-        buildZipFile(temporaryFolder.newFile("plugin.jar")) {
-          dir("META-INF") {
-            file("plugin.xml") {
-              perfectXmlBuilder.modify {
-                additionalContent = """
-                <xi:include 
-                  xmlns:xi="http://www.w3.org/2001/XInclude" 
-                  href="applicationServices.xml" 
-                  includeUnless="com.jetbrains.plugins.structure.mocks.optionalDependency"/>
-              """.trimIndent()
-              }
-            }
-            file("applicationServices.xml",optionalDependencyXml)
-          }
-        }
-      }
-      with(plugin.appContainerDescriptor.services) {
-        assertEquals(1, size)
-        val applicationService = first()
-        assertEquals("com.jetbrains.plugins.structure.mocks.ProjectService", applicationService.serviceImplementation)
-      }
-    }
-  }
-
-  @Test
-  fun `xinclude includeUnless with system property being set to true`() {
-    withSystemProperty("com.jetbrains.plugins.structure.mocks.optionalDependency", true) {
-      val plugin = buildPluginSuccess(emptyList()) {
-        buildZipFile(temporaryFolder.newFile("plugin.jar")) {
-          dir("META-INF") {
-            file("plugin.xml") {
-              perfectXmlBuilder.modify {
-                additionalContent = """
-                <xi:include 
-                  xmlns:xi="http://www.w3.org/2001/XInclude" 
-                  href="applicationServices.xml" 
-                  includeUnless="com.jetbrains.plugins.structure.mocks.optionalDependency"/>
-              """.trimIndent()
-              }
-            }
-            file(
-              "applicationServices.xml", optionalDependencyXml
-            )
-          }
-        }
-      }
-      assertEquals(0, plugin.appContainerDescriptor.services.size)
-    }
-  }
-
-  @Test
-  fun `xinclude includeUnless with system property being set`() {
-    withSystemProperty("com.jetbrains.plugins.structure.mocks.optionalDependency", true) {
-      val plugin = buildPluginSuccess(emptyList()) {
-        buildZipFile(temporaryFolder.newFile("plugin.jar")) {
-          dir("META-INF") {
-            file("plugin.xml") {
-              perfectXmlBuilder.modify {
-                depends += """<depends optional="true" config-file="optionalDependency.xml">Optional Dependency</depends>"""
-                additionalContent = """
-                <xi:include 
-                  xmlns:xi="http://www.w3.org/2001/XInclude" 
-                  href="optionalDependency.xml" 
-                  includeUnless="com.jetbrains.plugins.structure.mocks.optionalDependency"/>
-              """.trimIndent()
-              }
-            }
-            file("optionalDependency.xml", optionalDependencyXml)
-          }
-        }
-      }
-      val optionalDescriptor = plugin.optionalDescriptors.single()
-      assertEquals("optionalDependency.xml", optionalDescriptor.configurationFilePath)
-      assertEquals(PluginDependencyImpl("Optional Dependency", true, false), optionalDescriptor.dependency)
-
-      assert(plugin.extensions.isEmpty())
-    }
-  }
-
-  @Test
-  fun `xinclude includeIf with system property`() {
-    withSystemProperty("com.jetbrains.plugins.structure.mocks.optionalDependency", true) {
-      val plugin = buildPluginSuccess(emptyList()) {
-        buildZipFile(temporaryFolder.newFile("plugin.jar")) {
-          dir("META-INF") {
-            file("plugin.xml") {
-              perfectXmlBuilder.modify {
-                depends += """<depends optional="true" config-file="optionalDependency.xml">Optional Dependency</depends>"""
-                additionalContent = """
-                <xi:include 
-                  xmlns:xi="http://www.w3.org/2001/XInclude" 
-                  href="optionalDependency.xml" 
-                  includeIf="com.jetbrains.plugins.structure.mocks.optionalDependency"/>
-              """.trimIndent()
-              }
-            }
-            file("optionalDependency.xml", optionalDependencyXml)
-          }
-        }
-      }
-      val optionalDescriptor = plugin.optionalDescriptors.single()
-      assertEquals("optionalDependency.xml", optionalDescriptor.configurationFilePath)
-      assertEquals(PluginDependencyImpl("Optional Dependency", true, false), optionalDescriptor.dependency)
-
-      assert(plugin.extensions.isEmpty())
-    }
-  }
-
-  private val optionalDependencyXml = """
-    <idea-plugin>
-      <extensions defaultExtensionNs="com.intellij">
-        <applicationService serviceImplementation="com.jetbrains.plugins.structure.mocks.ProjectService"/>
-      </extensions>
-    </idea-plugin>
-    """.trimIndent()
-
-  private fun withSystemProperty(property: String, value: Boolean, block: () -> Unit) {
-    try {
-      System.setProperty(property, value.toString())
-      block()
-    } finally {
-      System.clearProperty(property)
-    }
-  }
-
   private fun checkPluginValues(plugin: IdePlugin, isDirectoryBasedPlugin: Boolean) {
     assertEquals("https://kotlinlang.org", plugin.url)
     assertEquals("Kotlin", plugin.pluginName)
